+++ src/autodq/vscode/__init__.py
@@ -5,7 +5,7 @@ from importlib.resources import as_file, files
 from pathlib import Path
 
 
-EXTENSION_VERSION = "0.2.6"
+EXTENSION_VERSION = "0.2.7"
 
 
 def extension_path() -> Path: